requirements.txt
numpy~=1.24.2
py-vox-io==0.1
